+++ bfg9000/tools/ar.py
@@ -39,14 +39,6 @@ class ArLinker(SimpleBuildCommand):
     def can_link(self, format, langs):
         return format == self.builder.object_format
 
-    @property
-    def _has_link_macros(self):
-        # We only need to define LIBFOO_EXPORTS/LIBFOO_STATIC macros on
-        # platforms that have different import/export rules for libraries. We
-        # approximate this by checking if the platform uses import libraries,
-        # and only define the macros if it does.
-        return self.env.target_platform.has_import_library
-
     def compile_options(self, step):
         options = opts.option_list()
         if self.builder.object_format != 'coff':
@@ -56,10 +48,7 @@ class ArLinker(SimpleBuildCommand):
 
     def forwarded_compile_options(self, step):
         options = opts.option_list()
-        if self._has_link_macros:
-            options.append(opts.define(
-                library_macro(step.name, 'static_library')
-            ))
+        options.append(opts.define(library_macro(step.name, 'static_library')))
         return options
 
     def flags(self, options, global_options=None, output=None, mode='normal'):