+++ examples/04_subdirs/include/detail/export.hpp
@@ -1,14 +1,16 @@
 #ifndef INC_DETAIL_EXPORT_HPP
 #define INC_DETAIL_EXPORT_HPP
 
-#if defined(_WIN32) && !defined(SUB_LIBLIBRARY_STATIC)
+#ifdef SUB_LIBLIBRARY_STATIC
+#  define LIB_PUBLIC
+#elif defined(_WIN32)
 #  ifdef SUB_LIBLIBRARY_EXPORTS
 #    define LIB_PUBLIC __declspec(dllexport)
 #  else
 #    define LIB_PUBLIC __declspec(dllimport)
 #  endif
 #else
-#  define LIB_PUBLIC
+#  define LIB_PUBLIC [[gnu::visibility("default")]]
 #endif
 
 #endif