+++ examples/10_pkg_config/include/hello.hpp
@@ -1,18 +1,20 @@
 #ifndef INC_HELLO_HPP
 #define INC_HELLO_HPP
 
-#if defined(_WIN32) && !defined(LIBHELLO_STATIC)
+#ifdef LIBHELLO_STATIC
+#  define LIBHELLO_PUBLIC
+#elif defined(_WIN32)
 #  ifdef LIBHELLO_EXPORTS
 #    define LIBHELLO_PUBLIC __declspec(dllexport)
 #  else
 #    define LIBHELLO_PUBLIC __declspec(dllimport)
 #  endif
 #else
-#  define LIBHELLO_PUBLIC
+#  define LIBHELLO_PUBLIC [[gnu::visibility("default")]]
 #endif
 
 namespace hello {
-  void LIBHELLO_PUBLIC say_hello();
+  LIBHELLO_PUBLIC void say_hello();
 }
 
 #endif