+++ examples/13_submodules/lib/library.cpp
@@ -3,6 +3,6 @@
 #include <iostream>
 #include <string>
 
-void LIB_PUBLIC hello(const char *name) {
+LIB_PUBLIC void hello(const char *name) {
   std::cout << "Hello, " << name << std::string(ENTHUSIASM, '!') << std::endl;
 }