+++ test/data/dual_use_library/outer.hpp
@@ -1,18 +1,20 @@
 #ifndef INC_OUTER_HPP
 #define INC_OUTER_HPP
 
-#if defined(_WIN32) && !defined(LIBOUTER_STATIC)
+#ifdef LIBOUTER_STATIC
+#  define LIBOUTER_PUBLIC
+#elif defined(_WIN32)
 #  ifdef LIBOUTER_EXPORTS
 #    define LIBOUTER_PUBLIC __declspec(dllexport)
 #  else
 #    define LIBOUTER_PUBLIC __declspec(dllimport)
 #  endif
 #else
-#  define LIBOUTER_PUBLIC
+#  define LIBOUTER_PUBLIC [[gnu::visibility("default")]]
 #endif
 
 namespace outer {
-  void LIBOUTER_PUBLIC hello();
+  LIBOUTER_PUBLIC void hello();
 }
 
 #endif