+++ test/data/install/shared_b.hpp
@@ -1,18 +1,20 @@
 #ifndef INC_SHARED_B_HPP
 #define INC_SHARED_B_HPP
 
-#if defined(_WIN32) && !defined(LIBSHARED_B_STATIC)
+#ifdef LIBSHARED_B_STATIC
+#  define SHARED_B_PUBLIC
+#elif defined(_WIN32)
 #  ifdef LIBSHARED_B_EXPORTS
 #    define SHARED_B_PUBLIC __declspec(dllexport)
 #  else
 #    define SHARED_B_PUBLIC __declspec(dllimport)
 #  endif
 #else
-#  define SHARED_B_PUBLIC
+#  define SHARED_B_PUBLIC [[gnu::visibility("default")]]
 #endif
 
 namespace shared_b {
-  void SHARED_B_PUBLIC hello();
+  SHARED_B_PUBLIC void hello();
 }
 
 #endif