+++ test/data/library_defs/library.hpp
@@ -1,16 +1,18 @@
 #ifndef INC_LIBRARY_HPP
 #define INC_LIBRARY_HPP
 
-#if defined(_WIN32) && !defined(LIBLIBRARY_STATIC)
+#ifdef LIBLIBRARY_STATIC
+#  define LIB_PUBLIC
+#elif defined(_WIN32)
 #  ifdef LIBLIBRARY_EXPORTS
 #    define LIB_PUBLIC __declspec(dllexport)
 #  else
 #    define LIB_PUBLIC __declspec(dllimport)
 #  endif
 #else
-#  define LIB_PUBLIC
+#  define LIB_PUBLIC [[gnu::visibility("default")]]
 #endif
 
-extern "C" void LIB_PUBLIC hello();
+extern "C" LIB_PUBLIC void hello();
 
 #endif