+++ test/data/pkg_config_system/src/inner.hpp
@@ -1,18 +1,20 @@
 #ifndef INC_INNER_HPP
 #define INC_INNER_HPP
 
-#if defined(_WIN32) && !defined(LIBINNER_STATIC)
+#ifdef LIBINNER_STATIC
+#  define LIBINNER_PUBLIC
+#elif defined(_WIN32)
 #  ifdef LIBINNER_EXPORTS
 #    define LIBINNER_PUBLIC __declspec(dllexport)
 #  else
 #    define LIBINNER_PUBLIC __declspec(dllimport)
 #  endif
 #else
-#  define LIBINNER_PUBLIC
+#  define LIBINNER_PUBLIC [[gnu::visibility("default")]]
 #endif
 
 namespace inner {
-  void LIBINNER_PUBLIC use_ogg();
+  LIBINNER_PUBLIC void use_ogg();
 }
 
 #endif